pyproject.toml
[tool.poetry]
name = "lambda"
version = "0.1.0"
description = ""
authors = ["hrfmmr <i@hrfmmr.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-api-python-client = "^2.88.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^1.0.0"
boto3 = "^1.26.151"
slack-sdk = "^3.21.3"

[tool.poetry.scripts]
main = "lambda.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
